sage/pyproject.toml
[project]
name = "sage"
version = "1.0.0"
description = "SAGE — Study & Ask with GLM-5 Engine"
requires-python = ">=3.11"

dependencies = [
    "langgraph>=0.2.0",
    "langchain-ollama>=0.1.0",
    "langchain-core>=0.2.0",
    "notebooklm-py[browser]>=0.3.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
sage = "main:app"
